refactor: move Column to column.go. impl Column.IsNullT()
