+++ column.go
@@ -2,9 +2,103 @@ package sqlla
 
 import (
 	"fmt"
+	"go/ast"
 	"sort"
 	"strings"
+
+	"github.com/Masterminds/goutils"
 )
+
+type Columns []Column
+
+type Column struct {
+	Field         *ast.Field
+	Name          string
+	MethodName    string
+	typeName      string
+	PkgName       string
+	baseTypeName  string
+	altTypeName   string
+	typeParameter string
+	TableName     string
+	IsPk          bool
+}
+
+func (c Column) HasUnderlyingType() bool {
+	return c.baseTypeName != c.typeName
+}
+
+func (c Column) TypeName() string {
+	tn := c.typeName
+	if c.typeParameter != "" {
+		return tn + "[" + c.typeParameter + "]"
+	}
+	return tn
+}
+
+func (c Column) BaseTypeName() string {
+	if c.typeParameter != "" {
+		return c.baseTypeName + "[" + c.typeParameter + "]"
+	}
+	return c.baseTypeName
+}
+
+func (c Column) AltTypeName() string {
+	if c.altTypeName == "" {
+		return ""
+	}
+	if c.typeParameter != "" {
+		return c.altTypeName + "[" + c.typeParameter + "]"
+	}
+	return c.altTypeName
+}
+
+func (c Column) IsNullT() bool {
+	return c.baseTypeName == "sql.Null" && c.typeParameter != ""
+}
+
+func (c Column) typeNameForExpr() string {
+	tname := c.BaseTypeName()
+	if atn := c.AltTypeName(); atn != "" {
+		tname = atn
+	}
+	tname = c.exprize(tname)
+	tname = goutils.Capitalize(tname)
+	return tname
+}
+
+func (c Column) ExprTypeName() string {
+	return "Expr" + c.typeNameForExpr()
+}
+
+func (c Column) exprize(s string) string {
+	s = strings.TrimPrefix(s, "sql.")
+	s = strings.TrimPrefix(s, "time.")
+	s = strings.TrimPrefix(s, "mysql.")
+	return s
+}
+
+func (c Column) ExprMultiTypeName() string {
+	return "ExprMulti" + c.typeNameForExpr()
+}
+
+func (c Column) ExprValueIdentifier() string {
+	if c.typeName != c.baseTypeName {
+		return c.baseTypeName + "(v)"
+	}
+	return "v"
+}
+
+func (c Column) String() string {
+	return c.Name
+}
+
+func (c Column) FieldName() string {
+	if len(c.Field.Names) > 0 {
+		return c.Field.Names[0].Name
+	}
+	return ""
+}
 
 type Where []Expr
